ref: separate properties (#3432)
* ref: separate properties

* ref: separate properties

* ref: separate properties

* ref: separate properties
